Split up completion tests too
